:recycle: Refactor search product name for api
